users/permissions.py
from rest_framework import permissions


class AdminOnly(permissions.BasePermission):

    def has_permission(self, request, view):
        if request.user.is_superuser:
            return True
        if request.user.role == 'admin':
            return True
        else:
            return False
